# spec/openapi.yaml
openapi: 3.1.0
info:
  title: I3X - Industrial Information Interface eXchange API
  description: |
    A common API for Contextualized Manufacturing Information Platforms (CMIPs) as defined in RFC 001.

    This specification enables portable application development across different manufacturing
    information platforms by providing a vendor-agnostic API specification.

    ## Key Concepts

    - **ElementId**: Platform-specific unique string identifier for any element
    - **Namespace**: Logical grouping of related types and instances
    - **ObjectType**: Schema definition for a class of objects
    - **ObjectInstance**: Actual equipment, sensors, or processes with attribute values
    - **VQT**: Value-Quality-Timestamp format for all data values
    - **Relationships**: Organizational (HasParent/HasChildren) and Composition (HasComponent/ComponentOf)

    ## maxDepth Parameter

    Controls recursion through HasComponent relationships:
    - `maxDepth=0`: Infinite recursion (include all nested composed elements)
    - `maxDepth=1`: No recursion (just this element's value) - default
    - `maxDepth=N`: Recurse up to N levels deep

    ## Quality Values

    - `Good`: Value is valid and current
    - `GoodNoData`: No data available but connection is good
    - `Bad`: Value is invalid or connection failed
    - `Uncertain`: Value quality cannot be determined

  version: 1.0.0
  contact:
    name: CESMII - The Smart Manufacturing Institute
    url: https://www.cesmii.org
...
servers:
...
paths:
...
  /namespaces:
    get:
      operationId: getNamespaces
      summary: List all namespaces
      description: |
        Returns an array of Namespaces registered in the CMIP.
        All Namespaces MUST have a Namespace URI to support follow-up queries.

        **RFC Reference:** 4.1.1
      tags:
        - Exploratory
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Namespace'
              example:
                - uri: "https://www.abelara.com/ns/equipment"
                  displayName: "Abelara Equipment"
                - uri: "https://www.isa.org/isa95"
                  displayName: "ISA-95"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ============================================================================
  # RFC 4.1.2/4.1.3 - Object Types
  # ============================================================================
  /objecttypes:
    get:
      operationId: getObjectTypes
      summary: List all object types
      description: |
        Returns an array of Type definitions registered in the CMIP.
        All Types MUST have an ElementId to support follow-up queries.
        The response MAY be filtered by NamespaceURI.

        **RFC Reference:** 4.1.3
      tags:
        - Exploratory
      parameters:
        - name: namespaceUri
          in: query
          description: Filter types by namespace URI
          required: false
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectType'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /relationshiptypes:
    get:
      operationId: getRelationshipTypes
      summary: List all relationship types
      description: |
        Returns an array of relationship type definitions registered in the CMIP.

        At minimum, implementations MUST support organizational hierarchy relationship types:
        - **HasParent/HasChildren**: Topological/organizational hierarchy

        Implementations MAY support composition relationship types:
        - **HasComponent/ComponentOf**: Data composition relationships

        **RFC Reference:** 4.1.4
      tags:
        - Exploratory
      parameters:
        - name: namespaceUri
          in: query
          description: Filter relationship types by namespace URI
          required: false
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RelationshipType'
              example:
                - elementId: "HasParent"
                  displayName: "Has Parent"
                  namespaceUri: "https://www.i3x.org/relationships"
                  reverseOf: "HasChildren"
                - elementId: "HasChildren"
                  displayName: "Has Children"
                  namespaceUri: "https://www.i3x.org/relationships"
                  reverseOf: "HasParent"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /objects:
    get:
      operationId: getObjects
      summary: List object instances
      description: |
        Returns an array of instance objects. Optionally filter by TypeId.

        The returned payload MUST include the metadata indicated in RFC 3.1.1:
        - displayName
        - parentId
        - hasChildren/isComposition
        - namespaceUri

        **RFC Reference:** 4.1.5
      tags:
        - Exploratory
      parameters:
        - name: typeId
          in: query
          description: Filter instances by object type ElementId
          required: false
          schema:
            type: string
        - name: includeMetadata
          in: query
          description: Include optional metadata (RFC 3.1.2)
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectInstance'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      summary: List all subscriptions
      description: |
        Returns a list of all active subscriptions for the authenticated client.

        Note: This endpoint is informational and not explicitly defined in RFC 001.
      tags:
        - Subscriptions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createSubscription
      summary: Create a subscription
      description: |
        Registers a client for a new subscription. This initial handshake allows
        the CMIP to allocate resources for a client.

        The response MUST include a Subscription ID that may be used for follow-up calls.

        Monitoring begins when objects are registered via the /register endpoint.

        **RFC Reference:** 4.2.3.1
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '200':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSubscriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /subscriptions/{subscriptionId}/register:
    post:
      operationId: registerMonitoredItems
      summary: Register monitored items
      description: |
        Registers the ElementIds the client wishes to subscribe to.

        Upon registration, the CMIP MUST begin publishing changed values.
        This method is additive - the client can add additional monitored items later.

        The maxDepth parameter controls recursion through HasComponent relationships
        when publishing updates for composition elements.

        **RFC Reference:** 4.2.3.2
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterItemsRequest'
      responses:
        '200':
          description: Items registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterItemsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Subscription or element not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /subscriptions/{subscriptionId}/stream:
    get:
      operationId: streamSubscription
      summary: Open SSE stream (streaming)
      description: |
        Opens a Server-Sent Events (SSE) stream for real-time data delivery.

        This implements streaming (At Most Once) delivery - the CMIP will publish
        messages as data becomes available, but provides no guarantee of delivery.

        When the stream is active, updates are delivered immediately rather than
        being queued for /sync retrieval.

        **RFC Reference:** 4.2.3.2 (streaming)
      tags:
        - Subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionIdPath'
        - name: mode
          in: query
          description: Delivery mode (streaming for SSE)
          required: false
          schema:
            type: string
            enum: [streaming]
            default: streaming
      responses:
        '200':
          description: SSE stream opened successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionUpdate'
              description: |
                Stream of JSON arrays, each containing subscription updates.
                Each line is a complete JSON array that can be parsed independently.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    Namespace:
      type: object
      description: A logical scope that groups related types, instances, and relationships
      required:
        - uri
        - displayName
      properties:
        uri:
          type: string
          format: uri
          description: Unique namespace URI
          example: "https://www.abelara.com/ns/equipment"
        displayName:
          type: string
          description: Human-readable namespace name
          example: "Abelara Equipment"
    ObjectType:
      type: object
      description: Schema definition for a class of objects (RFC 3.3)
      required:
        - elementId
        - displayName
        - namespaceUri
        - schema
      properties:
        elementId:
          type: string
          description: Unique string identifier for the type
          example: "MachineType"
        displayName:
          type: string
          description: Human-readable type name
          example: "Machine"
        namespaceUri:
          type: string
          format: uri
          description: Namespace URI this type belongs to
          example: "https://www.abelara.com/ns/equipment"
        schema:
          type: object
          description: JSON Schema definition for this object type
          additionalProperties: true
    RelationshipType:
      type: object
      description: Definition of a relationship between objects (RFC 3.2)
      required:
        - elementId
        - displayName
        - namespaceUri
        - reverseOf
      properties:
        elementId:
          type: string
          description: Unique string identifier for the relationship type
          example: "HasParent"
        displayName:
          type: string
          description: Human-readable relationship type name
          example: "Has Parent"
        namespaceUri:
          type: string
          format: uri
          description: Namespace URI for this relationship type
          example: "https://www.i3x.org/relationships"
        reverseOf:
          type: string
          description: ElementId of the inverse relationship type
          example: "HasChildren"
    ObjectInstance:
      type: object
      description: An actual equipment, sensor, or process with attribute values (RFC 3.1)
      required:
        - elementId
        - displayName
        - typeId
        - isComposition
        - namespaceUri
      properties:
        elementId:
          type: string
          description: Unique string identifier for this instance
          example: "machine-001"
        displayName:
          type: string
          description: Human-readable instance name
          example: "CNC Machine #1"
        typeId:
          type: string
          description: ElementId of the object type this instance is derived from
          example: "MachineType"
        parentId:
          type: string
          nullable: true
          description: ElementId of the parent object (if any)
          example: "production-line-a"
        isComposition:
          type: boolean
          description: True if this element has composed child objects (HasComponent)
          example: false
        namespaceUri:
          type: string
          format: uri
          description: Namespace URI for this instance
          example: "https://www.abelara.com/ns/equipment"
        relationships:
          type: object
          description: Relationships to other objects (when includeMetadata=true)
          additionalProperties: true
...
    CreateSubscriptionRequest:
      type: object
      description: Request body for creating a subscription
      properties: {}
      additionalProperties: true
    CreateSubscriptionResponse:
      type: object
      description: Response from creating a subscription
      required:
        - subscriptionId
        - message
      properties:
        subscriptionId:
          type: string
          description: Unique identifier for the new subscription
          example: "0"
        message:
          type: string
          description: Human-readable confirmation message
          example: "Subscription created successfully."
    SubscriptionList:
      type: object
      description: List of subscriptions
      required:
        - subscriptionIds
      properties:
        subscriptionIds:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionSummary'
    SubscriptionSummary:
      type: object
      description: Brief subscription information
      required:
        - subscriptionId
        - created
      properties:
        subscriptionId:
          type: string
          description: Unique subscription identifier
        created:
          type: string
          format: date-time
          description: Timestamp when subscription was created
...
    RegisterItemsRequest:
      type: object
      description: Request body for registering or unregistering monitored items
      required:
        - elementIds
      properties:
        elementIds:
          type: array
          items:
            type: string
          description: Array of ElementIds to monitor
          example: ["sensor-001", "sensor-002"]
        maxDepth:
          type: integer
          minimum: 0
          default: 1
          description: |
            Controls recursion through HasComponent relationships:
            - 0: Infinite recursion
            - 1: No recursion (default)
            - N: Recurse up to N levels
    RegisterItemsResponse:
      type: object
      description: Response from registering monitored items
      required:
        - message
        - totalObjects
      properties:
        message:
          type: string
          example: "Registered 5 objects to subscription."
        totalObjects:
          type: integer
          description: Total number of monitored objects after registration
...
    SubscriptionUpdate:
      type: object
      description: A single update from a subscription
      required:
        - elementId
        - value
      properties:
        elementId:
          type: string
          description: ElementId of the changed element
        value:
          description: The new value
        quality:
          type: string
          description: Quality indicator
          enum:
            - Good
            - GoodNoData
            - Bad
            - Uncertain
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the change

    # --------------------------------------------------------------------------
    # Error Response
    # --------------------------------------------------------------------------
    Error:
      type: object
      description: Standard error response
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Element 'unknown-id' not found"

  # ============================================================================
  # Parameters
  # ============================================================================
  parameters:
...
    SubscriptionIdPath:
      name: subscriptionId
      in: path
      required: true
      description: Unique subscription identifier
      schema:
        type: string
      example: "0"
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid request body"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Authentication required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Element 'unknown-id' not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Internal server error"

  # ============================================================================
  # Security Schemes
  # ============================================================================
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (RFC 5.3.1)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT/OAuth bearer token authentication (optional)
security:
  - ApiKeyAuth: []
  - BearerAuth: []